pyproject.toml
[project]
name = "mixed-iotester"
version = "0.1.0"
description = "Mixed IO workload generator with fixed mix scheduler"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

